requirements.txt
streamlit>=1.24.0
opencv-python
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0.0
requests>=2.31.0
plotly>=5.15.0
pyserial>=3.5 